+++ extensions/sql/transfer-process-store-sql/src/main/java/org/eclipse/dataspaceconnector/sql/transferprocess/store/SqlTransferProcessStore.java
@@ -36,7 +36,7 @@ import org.jetbrains.annotations.Nullable;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.time.Instant;
+import java.time.Clock;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
@@ -53,22 +53,23 @@ public class SqlTransferProcessStore implements TransferProcessStore {
     private final TransferProcessStoreStatements statements;
     private final String leaseHolderName;
     private final SqlLeaseContextBuilder leaseContext;
+    private final Clock clock;
 
-    public SqlTransferProcessStore(DataSourceRegistry dataSourceRegistry, String datasourceName, TransactionContext transactionContext, ObjectMapper objectMapper, TransferProcessStoreStatements statements, String leaseHolderName) {
-
+    public SqlTransferProcessStore(DataSourceRegistry dataSourceRegistry, String datasourceName, TransactionContext transactionContext, ObjectMapper objectMapper, TransferProcessStoreStatements statements, String leaseHolderName, Clock clock) {
         this.dataSourceRegistry = dataSourceRegistry;
         this.datasourceName = datasourceName;
         this.transactionContext = transactionContext;
         this.objectMapper = objectMapper;
         this.statements = statements;
         this.leaseHolderName = leaseHolderName;
-        leaseContext = SqlLeaseContextBuilder.with(transactionContext, leaseHolderName, statements);
+        this.clock = clock;
+        leaseContext = SqlLeaseContextBuilder.with(transactionContext, leaseHolderName, statements, clock);
     }
 
     @Override
     public @NotNull List<TransferProcess> nextForState(int state, int max) {
         var list = new ArrayList<TransferProcess>();
-        var now = Instant.now().toEpochMilli();
+        var now = clock.millis();
         transactionContext.execute(() -> {
             try (var conn = getConnection()) {
                 var stmt = statements.getNextForStateTemplate();