+++ spi/transfer-spi/src/test/java/org/eclipse/dataspaceconnector/spi/types/domain/transfer/TransferProcessTest.java
@@ -29,7 +29,6 @@ import java.util.UUID;
 import static java.util.Collections.emptyList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.eclipse.dataspaceconnector.spi.types.domain.transfer.TransferProcessStates.INITIAL;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -121,18 +120,6 @@ class TransferProcessTest {
         process.transitionDeprovisioning();
         process.transitionDeprovisioned();
         process.transitionEnded();
-    }
-
-    @Test
-    void verifyTransitionRollback() {
-        var process = TransferProcess.Builder.newInstance().id(UUID.randomUUID().toString()).build();
-        process.transitionInitial();
-        process.transitionProvisioning(ResourceManifest.Builder.newInstance().build());
-
-        process.rollbackState(INITIAL);
-
-        assertEquals(INITIAL.code(), process.getState());
-        assertEquals(1, process.getStateCount());
     }
 
     @Test