+++ CHANGELOG.md
@@ -25,7 +25,7 @@ in the detailed section referring to by linking pull requests or issues.
 
 #### Removed
 
-*
+* Old RegistrationService (was used for a PoC) (#164)
 
 #### Fixed
 
@@ -63,6 +63,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Verify OpenAPI definitions (#1312)
 * Documentation for CosmosDB (#1334)
 * Add validation to contract definition id (#1347)
+* Shared clock service (#1416)
 
 #### Changed